// src/modules/markdown-cleaner/core/types.ts
export interface CleanerOptions {
  removeHeading?: boolean;
  removeEmphasis?: boolean;
  removeStrong?: boolean;
  removeLink?: boolean;
  removeImage?: boolean;
  removeCode?: boolean;
  removeInlineCode?: boolean;
  removeBlockquote?: boolean;
  removeList?: boolean;
  removeTable?: boolean;
  removeThematicBreak?: boolean;
  removeDelete?: boolean;
  removeHtml?: boolean;
  removeBreak?: boolean;
  removeFootnote?: boolean;
}